style: refactor

+++ NineChess/src/graphicsconst.h
@@ -2,11 +2,11 @@
 #ifndef GRAPHICSCONST
 #define GRAPHICSCONST
 
-const int BOARD_SIZE = 600;     // 棋盘大小
-const int BOARD_MINISIZE = 150; // 最小宽高，即1/4大小
-const int PIECE_SIZE = 56;      // 棋子大小
-const int LINE_INTERVAL = 72;   // 线间距
-const int LINE_WEIGHT = 3;      // 线宽
+const short BOARD_SIZE = 600;     // 棋盘大小
+const short BOARD_MINISIZE = 150; // 最小宽高，即1/4大小
+const short PIECE_SIZE = 56;      // 棋子大小
+const short LINE_INTERVAL = 72;   // 线间距
+const short LINE_WEIGHT = 3;      // 线宽
 
 #endif // GRAPHICSCONST
